redix-core/requirements.txt
# Redix Green Intelligence Engine - Dependencies

# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
websockets==12.0

# LangChain and AI models
langchain==0.1.0
langchain-openai==0.0.5
langchain-anthropic==0.1.0
langchain-community==0.0.20
langgraph==0.0.20

# Local models
ollama==0.1.7

# System monitoring
psutil==5.9.8

# Caching and scaling
redis==5.0.1

# Data validation
pydantic==2.5.0

# Environment variables
python-dotenv==1.0.0

# HTTP client for Ollama checks
requests==2.31.0

